refactor(search): drop broken ADR-0010 persist-flag work, delete dead sort code, fix copyright/encapsulation
Per the detailed review from @yurishkuro:

- Removed the { persist: false } / _skipCount conditional-write mechanism
  and its "ADR-0010 PR 2" comment entirely. ADR-0010 is about TracePage
  layout settings, is still Status: Proposed, and its actual PR 2 was
  never implemented anywhere - this store had nothing to do with it. The
  mechanism was also broken by design: zustand's persist serializes the
  whole state blob on every write, so skipping one write doesn't make a
  single field ephemeral - the next persisted change of any other field
  re-includes it. Nothing in production called it, only the store's own
  tests did.
- Deleted model/search.ts and model/search.test.js (sortTraces,
  TraceOrderBy, ISortableTrace) - dead code with zero consumers once the
  model->component import keeping them alive was removed. Also removed
  the now-orphaned TraceOrderBy type from order-by.ts.
- Restored the original 2017 Uber Technologies copyright line (alongside
  2026 Jaeger Authors) on order-by.ts, sort.ts, and sort.test.ts - their
  content was extracted from pre-2026 files, not newly authored.
- setSortBy is now typed (sortBy: OrderBy) instead of string;
  sanitizeSortBy is module-private. SearchResults/index.tsx's
  handleSortChange no longer re-validates a value the store already
  validates - it just forwards to trackSortByChange directly.
- TraceTable now reads sortBy/setSortBy from the store directly, matching
  how it already reads startTimeDisplay, instead of taking them as props
  from a parent reading the same store.
- migrate is now the identity function plus a comment explaining why it
  can't be removed outright - merge already sanitizes every field on
  every rehydration path, including a version mismatch.
- store.search-results.ts (+ test) moved into SearchResults/, alongside
  its four consumers, for the same reason order-by.ts moved there.
- Removed the SearchResults/index.test.jsx test that fired an invalid
  sort key through a mocked dropdown to check component-level
  re-sanitization - that re-sanitization is gone, so the test exercised
  a state that's no longer reachable through the real UI.

Verified against a real running app: uploaded a 3-trace file, confirmed
column-click sort, the Sort: dropdown (including Most Errors, one of the
previously-broken keys), and that sortBy/viewMode both survive a full
page reload with no re-selection needed, with no console errors.
